feat: impeliment product controller and service

+++ internal/db/sqlc/schema/product.sql
@@ -2,7 +2,7 @@ CREATE TABLE products (
   id SERIAL PRIMARY KEY,
   name TEXT NOT NULL,
   description TEXT NOT NULL,
-  price NUMERIC(10,2) NOT NULL,
+  price BIGINT NOT NULL,
   is_active BOOLEAN DEFAULT TRUE NOT NULL,
   created_at TIMESTAMP DEFAULT now() NOT NULL
 );
